ура как-то работает Many to many!

+++ back-end/peoples_world/src/main/java/nsu/ypprpo/peoples_world/Requests/UsersHobbiesRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UsersHobbiesRequest {
     private Long user;
-    private Long hobby;
+    private Integer hobby;
 }
