+++ back-end/peoples_world/src/main/java/nsu/ypprpo/peoples_world/repository/HobbiesRepository.java
@@ -3,5 +3,5 @@ package nsu.ypprpo.peoples_world.repository;
 import nsu.ypprpo.peoples_world.models.Hobby;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HobbiesRepository extends JpaRepository<Hobby, Long> {
+public interface HobbiesRepository extends JpaRepository<Hobby, Integer> {
 }
